lint" (+4 squashed commits) Squashed commits: [a4d3c37] 100% ? [cdac8c6] update unit test [8369c56] 100% test coverage [1629533] split SupersetClient singleton and class into separate files
