+++ packages/superset-ui-connection/src/index.ts
@@ -1,3 +1,4 @@
 export { default as callApi } from './callApi';
 export { default as SupersetClient } from './SupersetClient';
+export { SupersetClientClass } from './SupersetClientClass';
 export * from './types';
